fix mapbox type

+++ src/main/kotlin/com/baulsupp/oksocial/services/mapbox/model/mapbox.kt
@@ -9,7 +9,7 @@ data class MapboxProperties(val address: String?, val category: String?, val tel
                             val landmark: Boolean?, val maki: String?)
 
 data class MapboxFeature(val id: String, val type: String, @Json(
-        name = "place_type") val placeType: List<String>, val relevance: Int,
+        name = "place_type") val placeType: List<String>, val relevance: Double,
                          val properties: MapboxProperties, val text: String, @Json(
                 name = "place_name") val placeName: String, val center: Location)
